+++ Artifacts/displacement.lua
@@ -13,9 +13,12 @@ local potentialEnvs = List.new() -- this is where we will put environments that 
 local validEnvs = List.new() -- this is where we will put all valid environments (no boar beach cuz its a secret and no risk of rain cuz its a final environment for example)
 	
 local hook = Hook.add_post(gm.constants.stage_roll_next, function(self, other, result, args)
+	print("hi")
 	if not displacement.active then return end
 	
 	validEnvs:clear()
+	
+	print("hi2")
 	
 	for i = 1, 4 do -- go through 4 tiers of stages, we exclude tier 6 because we dont want to randomly go to risk of rain, and we exclude tier 5 stages because artifact of tempus can easily be cheesed if temple of the elders is chosen as one of the first stages (it also breaks the onstep code of the artifact's object for some reason? no clue why)
 		local envs = List.wrap(gm._mod_stage_get_pool_list(i)) -- get all environments in a tier
@@ -40,6 +43,8 @@ local hook = Hook.add_post(gm.constants.stage_roll_next, function(self, other, r
 	
 	local randomEnv = math.random(1, #potentialEnvs) -- pick a random environment from list of potential environments
 	local upcomingEnv = potentialEnvs[randomEnv] -- set the next environment to the one picked above
+	
+	print(upcomingEnv)
 	
 	result.value = upcomingEnv -- make the script transport you to the upcoming environment instead of the usual one